fix: remaining -Wsign-compare and -Wmissing-designated-field-initializers in tests and metal2 API
- tensor.hpp: fix auto loop vars (deduced int) to uint32_t for shape comparisons
- bfloat_utils.hpp: change packed_data_size from int to size_t
- test_soc_descriptor.cpp: fix int vs size_t comparisons in device/core loops
- test_tilize_untilize.cpp: fix int vs uint32_t comparisons in tile/face loops
- kernel_spec.hpp: add = {} defaults to dfb_bindings, semaphore_bindings,
  tensor_bindings, compile_time_arg_bindings, CompilerOptions fields,
  RuntimeArgSchema named_runtime_args/named_common_runtime_args
- program_spec.hpp: add = {} defaults to all vector fields in ProgramSpec
- program_run_params.hpp: add = {} defaults to KernelRunParams vector/map fields

+++ tests/tt_metal/tt_metal/api/test_tilize_untilize.cpp
@@ -110,17 +110,17 @@ std::vector<T> convert_to_tile_layout(
     auto tile_HW = tile_H * tile_W;
     auto face_HW = face_H * face_W;
     TT_FATAL(data.size() % tile_HW == 0, "data size must be divisible by tile_HW");
-    int num_tiles = data.size() / tile_HW;
-    for (int tile_idx = 0; tile_idx < num_tiles; tile_idx++) {
+    uint32_t num_tiles = static_cast<uint32_t>(data.size() / tile_HW);
+    for (uint32_t tile_idx = 0; tile_idx < num_tiles; tile_idx++) {
         std::vector<T> top_left;
         std::vector<T> top_right;
         std::vector<T> bottom_left;
         std::vector<T> bottom_right;
 
         if (transpose_face) {
-            for (int col = 0; col < tile_W; col++) {
-                int index = (tile_idx * tile_HW) + col;
-                for (int row = 0; row < tile_H; row++) {
+            for (uint32_t col = 0; col < tile_W; col++) {
+                uint32_t index = (tile_idx * tile_HW) + col;
+                for (uint32_t row = 0; row < tile_H; row++) {
                     if (row < face_H and col < face_W) {
                         top_left.push_back(data[index]);
                     } else if (row < face_H and col >= face_W) {
@@ -136,9 +136,9 @@ std::vector<T> convert_to_tile_layout(
                 }
             }
         } else {
-            int index = tile_idx * tile_HW;
-            for (int row = 0; row < tile_H; row++) {
-                for (int col = 0; col < tile_W; col++) {
+            uint32_t index = tile_idx * tile_HW;
+            for (uint32_t row = 0; row < tile_H; row++) {
+                for (uint32_t col = 0; col < tile_W; col++) {
                     if (row < face_H and col < face_W) {
                         top_left.push_back(data[index]);
                     } else if (row < face_H and col >= face_W) {
